requirements.txt
anthropic>=0.39.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
notion-client>=2.2.1
